dataflow/generator/requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.1
diffusers @ git+https://github.com/huggingface/diffusers.git@99c0483b67427de467f11aa35d54678fd36a7ea2
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
einops==0.8.0
exceptiongroup==1.2.2
fastapi==0.115.2
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
gguf==0.10.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.2
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
importlib_metadata==8.5.0
interegular==0.3.3
Jinja2==3.1.4
jiter==0.6.1
jsonargparse==4.33.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lark==1.2.2
llvmlite==0.43.0
lm-format-enforcer==0.10.6
MarkupSafe==3.0.1
matplotlib==3.9.2
mistral_common==1.4.4
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
openai==1.51.2
opencv-python-headless==4.10.0.84
outlines==0.0.46
packaging==24.1
pandas==2.2.3
partial-json-parser==0.2.1.1.post4
pillow==10.4.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
propcache==0.2.0
protobuf==5.28.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==17.0.0
pycountry==24.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.37.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rpds-py==0.20.0
safetensors==0.4.5
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
starlette==0.40.0
style==1.1.0
sympy==1.13.3
tiktoken==0.7.0
tokenizers==0.20.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.45.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
update==0.0.1
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
vllm==0.6.3
watchfiles==0.24.0
websockets==13.1
xformers==0.0.27.post2
xxhash==3.5.0
yarl==1.15.3
zipp==3.20.2
decord==0.6.0
aisuite==0.1.10
docstring_parser==0.16